requirements.txt
aiofiles==23.2.1
aiohttp==3.8.3
httpx==0.27.0
numpy==1.24.1
Pillow==10.3.0
qq_botpy==1.2.0
